+++ square_common_bl/pydantic_models/internal.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class GetAppIdV0Response(BaseModel):
+    main: str